// src/app/domain/models/bitacoraAcceso.ts
export interface BitacoraAcceso {
  identificador: number;
  aplicativoId: string;
  capa: string;
  metodo: string;
  actividad: string;
  transaccionId: string;
  ipEquipo: string;
  fechaHoraAcceso: string;
  usuarioAcceso: string;
  nombreUsuario: string;
  expedienteUsuario: string;
  rfcUsuario: string;
  areaUsuario: string;
  puestoUsuario: string;
  estatusOperacion: string;
  respuestaOperacion: string;
  }