// src/app/domain/models/bitacoraUsuario.ts
export interface BitacoraUsuario {
  identificador: number;
  aplicativoId: string;
  capa: string;
  metodo: string;
  proceso: string;
  subproceso: string;
  detalleOperacion: string;
  transaccionId: string;
  ipEquipo: string;
  fechaHoraOperacion: string;
  usuarioOperador: string;
  nombreOperador: string;
  expedienteOperador: string;
  rfcOperador: string;
  areaOperador: string;
  puestoOperador: string;
  usuario: string;
  nombreUsuario: string;
  expedienteUsuario: string;
  rfcUsuario: string;
  areaUsuario: string;
  puestoUsuario: string;
  estatusOperacion: string;
  respuestaOperacion: string;
}